Fix syntax issues on Windows systems
The code uses pathlib to manipulate paths, and this apparently will take
a posix path and convert it into a local path. While this is OK for the
purposes of iterating the packages folder, it blocks syntaxes from
being able to load because package resources are always posix paths
regardless of the platform.

This change ensures that when working with the package paths we always
use a posix path.